# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sky Wings API",
        "description": "API documentation for Sky Wings application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookings": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "e33fa5a6507581f9b1e0d710a9a130d1",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "Authenticated user's ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "flight_id",
                                    "isBusiness",
                                    "passengers"
                                ],
                                "properties": {
                                    "flight_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "isBusiness": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "passengers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "surname": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                },
                                                "is_female": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "date_of_birth": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2000-01-01"
                                                },
                                                "seat_number": {
                                                    "type": "string",
                                                    "example": "12A"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "+123456789"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Booking created."
                                        },
                                        "data": {
                                            "type": "integer",
                                            "example": 101
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/bookings/my-flights": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get upcoming flights for authenticated user",
                "operationId": "dc2aa1bafb372afcdd084cbc3600e1c0",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "Authenticated user's ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of upcoming flights",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/bookings/history": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get booking history for authenticated user",
                "operationId": "3adb24dedf526c89fbf413cc54efd034",
                "parameters": [
                    {
                        "name": "user-id",
                        "in": "header",
                        "description": "Authenticated user's ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}